Fixed the bug where "Only the owner of the image can edit the image"

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -27,6 +27,7 @@ public class ImageController {
     @Autowired
     private TagService tagService;
 
+    HttpSession session;
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
     public String getUserImages(Model model) {
@@ -34,24 +35,6 @@ public class ImageController {
         model.addAttribute("images", images);
         return "images";
     }
-
-//    //This method is called when the details of the specific image with corresponding title are to be displayed
-//    //The logic is to get the image from the databse with corresponding title. After getting the image from the database the details are shown
-//    //First receive the dynamic parameter in the incoming request URL in a string variable 'title' and also the Model type object
-//    //Call the getImageByTitle() method in the business logic to fetch all the details of that image
-//    //Add the image in the Model type object with 'image' as the key
-//    //Return 'images/image.html' file
-//
-//    //Also now you need to add the tags of an image in the Model type object
-//    //Here a list of tags is added in the Model type object
-//    //this list is then sent to 'images/image.html' file and the tags are displayed
-//    @RequestMapping("/images/{title}")
-//    public String showImage(@PathVariable("title") String title, Model model) {
-//        Image image = imageService.getImageByTitle(title);
-//        model.addAttribute("image", image);
-//        model.addAttribute("tags", image.getTags());
-//        return "images/image";
-//    }
 
     //This method is called when the details of the specific image with corresponding title are to be displayed
     //The logic is to get the image from the databse with corresponding title. After getting the image from the database the details are shown
@@ -111,12 +94,24 @@ public class ImageController {
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
     public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+        String error = "Only the owner of the image can edit the image";
         Image image = imageService.getImage(imageId);
+        User owner = image.getUser();
+        String creatorUsername = owner.getUsername();
+        User user = (User) session.getAttribute("loggeduser");
 
+        if(!creatorUsername.equalsIgnoreCase(user.getUsername())){
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+        model.addAttribute("editError", error);
+        return "images/image";
+        }else{
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -158,11 +153,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model) {
+        String error = "Only the owner of the image can delete the image";
+        User user = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        User owner = image.getUser();
+        String creatorUsername = owner.getUsername();
+        if(!creatorUsername.equalsIgnoreCase(user.getUsername())){
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }else {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
-
 
     //This method converts the image to Base64 format
     private String convertUploadedFileToBase64(MultipartFile file) throws IOException {
